project/backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
mysql-connector-python==8.0.32
python-dotenv==1.0.0
numpy==1.24.2
pandas==1.5.3
scikit-learn==1.2.2